Change request live date logic
